+++ up42/templates.py
@@ -1,9 +1,10 @@
 import dataclasses
-from typing import Union
+from typing import List, Optional, Union
 
 from up42 import base, processing
 
 
+# TODO: drop these with Python 3.10 kw_only=True data classes
 @dataclasses.dataclass
 class WorkspaceIdSingleItemTemplate(processing.SingleItemJobTemplate):
     workspace_id: Union[str, base.WorkspaceId] = dataclasses.field(default=base.WorkspaceId())
@@ -77,3 +78,52 @@ class HypervergePleiadesChangeDetection(WorkspaceIdMultiItemTemplate):
 @dataclasses.dataclass
 class HypervergeSpotChangeDetection(WorkspaceIdMultiItemTemplate):
     process_id = "detection-change-spot-hyperverge"
+
+
+@dataclasses.dataclass
+class SpaceptAugmentation(processing.SingleItemJobTemplate):
+    denoising_factor: int = 0
+    colour_denoising_factor: int = 10
+    workspace_id: Union[str, base.WorkspaceId] = dataclasses.field(default=base.WorkspaceId())
+
+    @property
+    def inputs(self):
+        return {
+            **super().inputs,
+            "denoising_factor": self.denoising_factor,
+            "colour_denoising_factor": self.colour_denoising_factor,
+        }
+
+
+@dataclasses.dataclass
+class NSUpsamling(processing.SingleItemJobTemplate):
+    ned: bool = False
+    rgb: bool = True
+    workspace_id: Union[str, base.WorkspaceId] = dataclasses.field(default=base.WorkspaceId())
+
+    @property
+    def inputs(self):
+        return {
+            **super().inputs,
+            "NED": self.ned,
+            "RGB": self.rgb,
+        }
+
+
+@dataclasses.dataclass
+class GreyWeight:
+    band: str
+    weight: float
+
+
+@dataclasses.dataclass
+class Pansharpening(processing.SingleItemJobTemplate):
+    grey_weights: Optional[List[GreyWeight]] = None
+    workspace_id: Union[str, base.WorkspaceId] = dataclasses.field(default=base.WorkspaceId())
+
+    @property
+    def inputs(self):
+        weights = (
+            {"greyWeights": [dataclasses.asdict(weight) for weight in self.grey_weights]} if self.grey_weights else {}
+        )
+        return {**super().inputs, **weights}