+++ CHANGELOG.md
@@ -29,11 +29,34 @@ You can check your current version with the following command:
     ```
 
 For more information, see [UP42 Python package description](https://pypi.org/project/up42-py/).
+## 1.0.4a15
+
+**Jun 11, 2024**
+
+- Added job template execution to `templates.py`
+
+# 1.0.4a14
+
+**Jun 10, 2024**
+
+- Added class `Job` to `processing` module to access processing job features.
+
+## 1.0.4a13
+
+**June 10, 2024**
+- Allow dependency `geopandas` from 0.13.2 to any version less than `1`.
+
+## 1.0.4a12
+
+**Jun 10, 2024**
+
+- Bumped dependencies `tornado` from 6.4 to 6.4.1.
+
 ## 1.0.4a11
 
 **Jun 10, 2024**
 
-- Added job template execution to `templates.py`
+- Added job templates for `SpaceptAugmentation`, `NSUpsamling`, `Pansharpening` to `templates.py`
 
 ## 1.0.4a10